move helper methods to util for later reports
